[refact]: operator norm 수정
