+++ operator/vec_operator3.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   vec_operator3.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jonhan <jonhan@student.42.fr>              +#+  +:+       +#+        */
+/*   By: seojchoi <seojchoi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/22 14:21:36 by jonhan            #+#    #+#             */
-/*   Updated: 2023/10/22 14:22:04 by jonhan           ###   ########.fr       */
+/*   Updated: 2023/11/20 15:15:54 by seojchoi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,4 +39,23 @@ t_vec	vec_min(t_vec v1, t_vec v2)
 	else
 		new.z = v2.z;
 	return (new);
+}
+
+t_vec	ray_at(t_ray ray, double t)
+{
+	t_vec	out;
+
+	out.x = ray.origin.x + ray.dir.x * t;
+	out.y = ray.origin.y + ray.dir.y * t;
+	out.z = ray.origin.z + ray.dir.z * t;
+	return (out);
+}
+
+t_ray	ray(t_vec origin, t_vec dir)
+{
+	t_ray	r;
+
+	r.origin = origin;
+	r.dir = dir;
+	return (r);
 }
