--- ray.c
@@ -1,22 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   ray.c                                              :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: jonhan <jonhan@student.42.fr>              +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/22 15:12:23 by jonhan            #+#    #+#             */
-/*   Updated: 2023/10/22 15:12:43 by jonhan           ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "miniRT.h"
-
-t_ray	ray(t_vec origin, t_vec dir)
-{
-	t_ray	r;
-
-	r.origin = origin;
-	r.dir = dir;
-	return (r);
-}
